- Merging IO to /Extensions.cs and /Util.cs - Deleted LocalFileSettingsNoVersionProvider, LocalFileSettingsProvider is going to set IgnoreVersion to true by default - Renamed LogType LogTypes - Updated LogSeverity member values so they are no longer bit flags - Merged sub namespaces in /Data, renamed *Impl to *Sql*Helper

+++ Cryptography/KeyFormat.cs
@@ -5,7 +5,7 @@ namespace Adenson.Cryptography
 	/// <summary>
 	/// Types of crypto keys
 	/// </summary>
-	public enum KeyFormat : byte
+	public enum KeyFormat : int
 	{
 		/// <summary>
 		/// Key is stored in base 64
